refactor: move initRoute to route.ts
